tech(api): move assessmentRepository
