+++ api/lib/application/preHandlers/assessment-authorization.js
@@ -1,4 +1,4 @@
-import * as assessmentRepository from '../../infrastructure/repositories/assessment-repository.js';
+import * as assessmentRepository from '../../../src/shared/infrastructure/repositories/assessment-repository.js';
 import * as validationErrorSerializer from '../../infrastructure/serializers/jsonapi/validation-error-serializer.js';
 import * as requestResponseUtils from '../../infrastructure/utils/request-response-utils.js';
