+++ api/lib/domain/services/placement-profile-service.js
@@ -3,7 +3,7 @@ import bluebird from 'bluebird';
 
 import { UserCompetence } from '../models/UserCompetence.js';
 import { PlacementProfile } from '../models/PlacementProfile.js';
-import * as assessmentRepository from '../../infrastructure/repositories/assessment-repository.js';
+import * as assessmentRepository from '../../../src/shared/infrastructure/repositories/assessment-repository.js';
 import * as skillRepository from '../../infrastructure/repositories/skill-repository.js';
 import * as assessmentResultRepository from '../../infrastructure/repositories/assessment-result-repository.js';
 import * as knowledgeElementRepository from '../../infrastructure/repositories/knowledge-element-repository.js';
