+++ api/tests/integration/domain/usecases/create-mission-assessment_test.js
@@ -1,6 +1,6 @@
 import _ from 'lodash';
 import { expect, knex } from '../../../test-helper.js';
-import * as assessmentRepository from '../../../../lib/infrastructure/repositories/assessment-repository.js';
+import * as assessmentRepository from '../../../../src/shared/infrastructure/repositories/assessment-repository.js';
 import { createMissionAssessment } from '../../../../lib/domain/usecases/create-mission-assessment.js';
 
 describe('Integration | UseCases | create-mission-assessment', function () {
