+++ api/tests/integration/domain/usecases/create-preview-assessment_test.js
@@ -1,6 +1,6 @@
 import { expect, knex } from '../../../test-helper.js';
 import { createPreviewAssessment } from '../../../../lib/domain/usecases/create-preview-assessment.js';
-import * as assessmentRepository from '../../../../lib/infrastructure/repositories/assessment-repository.js';
+import * as assessmentRepository from '../../../../src/shared/infrastructure/repositories/assessment-repository.js';
 import { Assessment } from '../../../../lib/domain/models/index.js';
 
 describe('Integration | UseCases | create-preview-assessment', function () {
