+++ api/tests/integration/domain/usecases/improve-competence-evaluation_test.js
@@ -2,7 +2,7 @@ import { expect, databaseBuilder, knex } from '../../../test-helper.js';
 import { improveCompetenceEvaluation } from '../../../../lib/domain/usecases/improve-competence-evaluation.js';
 
 import * as competenceEvaluationRepository from '../../../../lib/infrastructure/repositories/competence-evaluation-repository.js';
-import * as assessmentRepository from '../../../../lib/infrastructure/repositories/assessment-repository.js';
+import * as assessmentRepository from '../../../../src/shared/infrastructure/repositories/assessment-repository.js';
 import { getCompetenceLevel } from '../../../../lib/domain/services/get-competence-level.js';
 
 describe('Integration | UseCase | Improve Competence Evaluation', function () {
